+++ src/client/RouteSelector/RouteSelector.story.tsx
@@ -1,32 +1,14 @@
 import { action } from "@storybook/addon-actions";
 import { boolean } from "@storybook/addon-knobs";
 import { storiesOf } from "@storybook/react";
-import { Button, Form } from "antd";
 import * as React from "react";
 
 import { Airport } from "../Airport";
 import { Route } from "../Route";
-import { RouteSelector, RouteSelectorProps } from "./RouteSelector";
+import { withForm } from "../withForm";
+import { RouteSelectorWrapped as RouteSelector } from "./RouteSelector";
 
-const RouteSelectorWrapped = Form.create()((props: RouteSelectorProps) => {
-  const onSubmit = (event: React.FormEvent<any>) => {
-    event.preventDefault();
-
-    props.form.validateFields();
-  };
-
-  return (
-    <Form onSubmit={onSubmit}>
-      <RouteSelector {...props} />
-      <Button
-        type="primary"
-        htmlType="submit"
-      >
-        Submit
-      </Button>
-    </Form>
-  );
-});
+const RouteSelectorWrapped = withForm(RouteSelector);
 
 storiesOf("RouteSelector", module)
   .add("default", () => (
@@ -35,6 +17,15 @@ storiesOf("RouteSelector", module)
         new Airport("WRO", "Wrocław", ["DUB"]),
         new Airport("DUB", "Dublin", ["WRO"]),
       ]}
+      isRequired={boolean("Required", true)}
+      isDisabled={boolean("Disabled", false)}
+      value={new Route()}
+      onChange={action("onChange")}
+    />
+  ))
+  .add("empty", () => (
+    <RouteSelectorWrapped
+      airports={[]}
       isRequired={boolean("Required", true)}
       isDisabled={boolean("Disabled", false)}
       value={new Route()}
